raster-nodes: create `mod blending_nodes` and move assoc nodes
